fix: Clean up of text_files_equal (#664)
